backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
requests==2.31.0 
sqlalchemy==2.0.36
pymysql==1.1.0
pillow
pytesseract==0.3.13
pandas
numpy
pydantic
opencv-python